GA fix, added heuristics

+++ code/src/main/java/hr/fer/zavrad/data/Data.java
@@ -1,3 +1,3 @@
 package hr.fer.zavrad.data;
 
-public record Data(String name, int capacity, int n, int solution, double[] items) {}
+public record Data(String name, double capacity, int n, int solution, double[] items) {}
